Presenter classes were refactored
