--- app/src/main/java/com/pokidin/a/diary/contracts/CheckInContract.java
@@ -1,32 +1,0 @@
-package com.pokidin.a.diary.contracts;
-
-import com.pokidin.a.diary.common.UserData;
-
-public interface CheckInContract {
-
-//    interface CheckInModel {
-//        void sendCheckInUserData(UserData userData);
-//
-//    }
-
-//    interface CheckInView {
-//        void initView();
-//
-//        UserData getUserData();
-//
-//        void showToast(String message);
-//
-//        void showProgress();
-//
-//        void hideProgress();
-//    }
-//
-//    interface CheckInPresenter {
-//
-//        void checkInBtnClicked();
-//
-//        void checkInUser();
-//
-//        void onDestroy();
-//    }
-}
